Add test documents from tika test suite
This patch adds a zip of about 200 files from tika's test suite,
and we assert some content comes back from each. This is a good exercise
of the various formats.

I removed any huge files to try to keep size reasonable, but we want
a bit of a variety so we know stuff is working.

I fixed issues with the parser config by running this.


Former-commit-id: 6b91e53ab587314faa64e324a11a0f09b393a15b