Introduce a relocated state during primary relocation
* this is a temporary fix until a more permanent fix is done on master *

During primary relocation, some operation can be done on the source primary but reach the target primary only after the relocation is completed. At the moment the new primary will have a new primary term and as such it will reject the operations from the old one, causing data loss.

This changes relocations to move the source primary to a relocated state, prevent any new operations from happening on it and waits for ongoing operations to complete.

Long we term we may also conisder not incrementing the primary term on relocation.


Former-commit-id: 3d061a8e6259006b0a849b0a7f059783af25b9f6