ingest: Added foreach processor.
This processor is useful when all elements of a json array need to be processed in the same way.
This avoids that a processor needs to be defined for each element in an array.
Also it is very likely that it is unknown how many elements are inside an json array.


Former-commit-id: 7a6adfd93ad4e41663642124e29993e14b6f19df